Add initial project files and structure
This commit adds the initial project files and structure, including the .gitignore file, VSCode configuration file, constants for assets images and router, and utility function for generating routes. It also includes the SplashView and OnBoardingView widgets, as well as their respective components. Finally, it updates the main.dart file to include the necessary dependencies and initial routes.